Working CUDA implementation

+++ src/cuda/kernels/skybox.cuh
@@ -2,6 +2,6 @@
 
 #include "color.cuh"
 
-__device__ Color get_skybox_color(const double3 &direction) {
+__device__ Color get_skybox_color(const float3 &direction) {
     return Color();
 }
